+++ datadog_checks_base/datadog_checks/base/checks/windows/perf_counters/base.py
@@ -1,8 +1,6 @@
 # (C) Datadog, Inc. 2021-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-# TODO: when we stop invoking Mypy with --py2, remove ignore and use f-strings
-# type: ignore
 from collections import ChainMap
 from contextlib import contextmanager, suppress
 
@@ -74,7 +72,7 @@ class PerfCountersBaseCheck(AgentCheck):
             # https://mhammond.github.io/pywin32/win32pdh__CollectQueryData_meth.html
             win32pdh.CollectQueryData(self._connection.query_handle)
         except pywintypes.error as error:
-            message = 'Error querying performance counters: {}'.format(error.strerror)
+            message = f'Error querying performance counters: {error.strerror}'
             self.submit_health_check(self.CRITICAL, message=message)
             self.log.error(message)
             return
@@ -93,19 +91,19 @@ class PerfCountersBaseCheck(AgentCheck):
         config = self.get_config_with_defaults()
         server_tag = config.get('server_tag') or 'server'
         use_localized_counters = is_affirmative(self.init_config.get('use_localized_counters', False))
-        tags = ['{}:{}'.format(server_tag, self._connection.server)]
+        tags = [f'{server_tag}:{self._connection.server}']
         tags.extend(config.get('tags', []))
         self._static_tags = tuple(tags)
 
         for option_name in ('metrics', 'extra_metrics'):
             metric_config = config.get(option_name, {})
             if not isinstance(metric_config, dict):
-                raise ConfigTypeError('Setting `{}` must be a mapping'.format(option_name))
+                raise ConfigTypeError(f'Setting `{option_name}` must be a mapping')
 
             for object_name, object_config in metric_config.items():
                 if not isinstance(object_config, dict):
                     raise ConfigTypeError(
-                        'Performance object `{}` in setting `{}` must be a mapping'.format(object_name, option_name)
+                        f'Performance object `{object_name}` in setting `{option_name}` must be a mapping'
                     )
 
                 perf_object = self.get_perf_object(
@@ -184,12 +182,12 @@ class PerfCountersBaseCheckWithLegacySupport(PerfCountersBaseCheck):
                 metrics_config = updated_config['metrics'] = {}
                 for object_name, config in default_config['metrics'].items():
                     new_config = config.copy()
-                    new_config['name'] = '{}.{}'.format(self.__NAMESPACE__, new_config['name'])
+                    new_config['name'] = f'{self.__NAMESPACE__}.{new_config["name"]}'
                     metrics_config[object_name] = new_config
 
             # Ensure idempotency in case this method is called multiple times due to configuration errors
             if self.namespace:
-                self.SERVICE_CHECK_HEALTH = '{}.{}'.format(self.__NAMESPACE__, self.SERVICE_CHECK_HEALTH)
+                self.SERVICE_CHECK_HEALTH = f'{self.__NAMESPACE__}.{self.SERVICE_CHECK_HEALTH}'
 
             self.namespace = ''