+++ sqlserver/datadog_checks/sqlserver/statements.py
@@ -153,12 +153,7 @@ def _row_key(row):
     :param row: a normalized row from STATEMENT_METRICS_QUERY
     :return: a tuple uniquely identifying this row
     """
-    return (
-        row.get('database_name'),
-        row['query_signature'],
-        row['query_hash'],
-        row['query_plan_hash'],
-    )
+    return (row.get('database_name'), row['query_signature'], row['query_hash'], row.get('procedure_name'))
 
 
 XML_PLAN_OBFUSCATION_ATTRS = frozenset(
@@ -444,6 +439,7 @@ class SqlserverStatementMetrics(DBMAsyncJob):
             'sqlserver_engine_edition': self._check.static_info_cache.get(STATIC_INFO_ENGINE_EDITION, ""),
             'ddagentversion': datadog_agent.get_version(),
             'ddagenthostname': self._check.agent_hostname,
+            'service': self._config.service,
         }
 
     @tracked_method(agent_check_getter=agent_check_getter)
@@ -500,6 +496,7 @@ class SqlserverStatementMetrics(DBMAsyncJob):
                 "ddsource": "sqlserver",
                 "ddtags": ",".join(tags),
                 "dbm_type": "fqt",
+                'service': self._config.service,
                 "db": {
                     "instance": row.get('database_name', None),
                     "query_signature": row['query_signature'],
@@ -593,6 +590,7 @@ class SqlserverStatementMetrics(DBMAsyncJob):
                     "cloud_metadata": self._config.cloud_metadata,
                     'sqlserver_version': self._check.static_info_cache.get(STATIC_INFO_VERSION, ""),
                     'sqlserver_engine_edition': self._check.static_info_cache.get(STATIC_INFO_ENGINE_EDITION, ""),
+                    'service': self._config.service,
                     "db": {
                         "instance": row.get("database_name", None),
                         "plan": {