+++ sqlserver/tests/test_integration.py
@@ -382,9 +382,7 @@ def test_autodiscovery_multiple_instances(aggregator, dd_run_check, instance_aut
     found_log = 0
     for _, _, message in caplog.record_tuples:
         # make sure master and msdb is only queried once
-        if "SqlDatabaseFileStats: changing cursor context via use statement: use [master]" in message:
-            found_log += 1
-        if "SqlDatabaseFileStats: changing cursor context via use statement: use [msdb]" in message:
+        if "Restoring the original database context master" in message:
             found_log += 1
 
     assert found_log == 2
@@ -813,6 +811,21 @@ def test_index_usage_statistics(aggregator, dd_run_check, instance_docker, datab
         aggregator.assert_metric(m, tags=expected_tags, count=1)
 
 
+@pytest.mark.integration
+@pytest.mark.usefixtures('dd_environment')
+def test_database_state(aggregator, dd_run_check, init_config, instance_docker):
+    instance_docker['database'] = 'master'
+    sqlserver_check = SQLServer(CHECK_NAME, init_config, [instance_docker])
+    dd_run_check(sqlserver_check)
+    expected_tags = sqlserver_check._config.tags + [
+        'database_recovery_model_desc:SIMPLE',
+        'database_state_desc:ONLINE',
+        'database:{}'.format(instance_docker['database']),
+        'db:{}'.format(instance_docker['database']),
+    ]
+    aggregator.assert_metric('sqlserver.database.state', tags=expected_tags, hostname=sqlserver_check.resolved_hostname)
+
+
 @pytest.mark.parametrize(
     'instance_propagate_agent_tags,init_config_propagate_agent_tags,should_propagate_agent_tags',
     [
@@ -860,3 +873,27 @@ def test_propagate_agent_tags(
                 status=SQLServer.OK,
                 tags=expected_tags,
             )
+
+
+@pytest.mark.integration
+@pytest.mark.usefixtures('dd_environment')
+def test_check_static_information_expire(aggregator, dd_run_check, init_config, instance_docker):
+    sqlserver_check = SQLServer(CHECK_NAME, init_config, [instance_docker])
+    dd_run_check(sqlserver_check)
+    assert sqlserver_check.static_info_cache is not None
+    assert len(sqlserver_check.static_info_cache.keys()) == 4
+    assert sqlserver_check.resolved_hostname == 'stubbed.hostname'
+
+    # manually clear static information cache
+    sqlserver_check.static_info_cache.clear()
+    dd_run_check(sqlserver_check)
+    assert sqlserver_check.static_info_cache is not None
+    assert len(sqlserver_check.static_info_cache.keys()) == 4
+    assert sqlserver_check.resolved_hostname == 'stubbed.hostname'
+
+    # manually pop STATIC_INFO_ENGINE_EDITION to make sure it is reloaded
+    sqlserver_check.static_info_cache.pop(STATIC_INFO_ENGINE_EDITION)
+    dd_run_check(sqlserver_check)
+    assert sqlserver_check.static_info_cache is not None
+    assert len(sqlserver_check.static_info_cache.keys()) == 4
+    assert sqlserver_check.resolved_hostname == 'stubbed.hostname'